move the tests to oracle and leave a separate crate for the utils
