+++ contracts/demo-issuer/Cargo.toml
@@ -5,7 +5,7 @@ authors = ["Robert Zaremba 'https://zaremba.ch/'"]
 edition = "2021"
 
 [lib]
-crate-type = ["cdylib", "lib"]
+crate-type = ["cdylib", "rlib"]
 
 
 [dependencies]
