+++ contracts/oracle/tests/integration_test.rs
@@ -2,8 +2,13 @@ use near_sdk::ONE_NEAR;
 use serde_json::json;
 use workspaces::{Account, AccountId, Contract, DevNetwork, Worker};
 
-use oracle_sbt::MINT_TOTAL_COST;
 use sbt::ContractMetadata;
+
+use oracle_sbt::{MINT_TOTAL_COST, MINT_TOTAL_COST_WITH_KYC};
+use std::str::FromStr;
+use test_util::common::ExternalAccountId;
+use test_util::utils::{build_signed_claim, generate_keys};
+use test_util::workspaces::{build_contract, gen_user_account};
 
 const AUTHORITY_KEY: &str = "zqMwV9fTRoBOLXwt1mHxBAF3d0Rh9E9xwSAXR3/KL5E=";
 const CLAIM_TTL: u64 = 3600 * 24 * 365 * 100;
@@ -111,6 +116,127 @@ async fn check_arithmetic_exception_mainnet() -> anyhow::Result<()> {
     assert!(tx.is_success(), "transfer: {:?}\n", tx.outcomes());
 
     check_arithmetic_exception(oracle, alice).await?;
+
+    Ok(())
+}
+
+#[tokio::test]
+async fn test_mint_sbt() -> anyhow::Result<()> {
+    let worker = workspaces::sandbox().await?;
+
+    let (sec_key, pub_key) = generate_keys();
+
+    let admin_account = gen_user_account(&worker, "admin.test.near").await?;
+
+    let registry_contract = build_contract(
+        &worker,
+        "../registry/",
+        json!({
+            "authority": admin_account.id(),
+        }),
+    )
+    .await?;
+
+    let oracle_contract = build_contract(
+        &worker,
+        "../oracle/",
+        json!({
+            "authority": near_sdk::base64::encode(pub_key.unwrap_as_ed25519().as_ref()),
+            "metadata": {
+                "spec": "v1.0.0",
+                "name": "test-sbt",
+                "symbol": "SBT"
+            },
+            "registry": registry_contract.id(),
+            "claim_ttl": 100000000000u64,
+            "admin": admin_account.id(),
+        }),
+    )
+    .await?;
+
+    let user_account = gen_user_account(&worker, "user.test.near").await?;
+    let signed_claim = build_signed_claim(
+        near_sdk::AccountId::from_str(user_account.id().as_str())?,
+        ExternalAccountId::gen(),
+        false,
+        &sec_key,
+    )?;
+
+    // TODO: add check for specific error text
+    let _ = user_account
+        .call(oracle_contract.id(), "sbt_mint")
+        .args_json(signed_claim)
+        .max_gas()
+        .transact()
+        .await?
+        .into_result()
+        .expect_err("only root and implicit accounts are allowed to get SBT");
+
+    let user_account = worker.root_account()?;
+    let signed_claim = build_signed_claim(
+        near_sdk::AccountId::from_str(user_account.id().as_str())?,
+        ExternalAccountId::gen(),
+        false,
+        &sec_key,
+    )?;
+
+    // TODO: add check for specific error text
+    let _ = user_account
+        .call(oracle_contract.id(), "sbt_mint")
+        .args_json(signed_claim)
+        .max_gas()
+        .transact()
+        .await?
+        .into_result()
+        .expect_err("Requires attached deposit of exactly 0.008 NEAR");
+
+    let signed_claim = build_signed_claim(
+        near_sdk::AccountId::from_str(user_account.id().as_str())?,
+        ExternalAccountId::gen(),
+        true,
+        &sec_key,
+    )?;
+
+    // TODO: add check for specific error text
+    let _ = user_account
+        .call(oracle_contract.id(), "sbt_mint")
+        .args_json(&signed_claim)
+        .max_gas()
+        .transact()
+        .await?
+        .into_result()
+        .expect_err("Requires attached deposit of exactly 0.015 NEAR");
+
+    // TODO: add check for specific error text
+    let _ = user_account
+        .call(oracle_contract.id(), "sbt_mint")
+        .args_json(json!({
+            "claim_b64": signed_claim.claim_b64,
+            "claim_sig": format!("a{}", &signed_claim.claim_sig),
+        }))
+        .deposit(MINT_TOTAL_COST_WITH_KYC)
+        .max_gas()
+        .transact()
+        .await?
+        .into_result()
+        .expect_err("can't base64-decode claim_sig");
+
+    let user_account = worker.root_account()?;
+    let signed_claim = build_signed_claim(
+        near_sdk::AccountId::from_str(user_account.id().as_str())?,
+        ExternalAccountId::gen(),
+        false,
+        &sec_key,
+    )?;
+
+    let _ = user_account
+        .call(oracle_contract.id(), "sbt_mint")
+        .args_json(signed_claim)
+        .deposit(MINT_TOTAL_COST)
+        .max_gas()
+        .transact()
+        .await?
+        .into_result()?;
 
     Ok(())
 }
